# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Templarc",
...
    "contact": {
      "name": "Templarc"
    },
    "license": {
      "name": "Private"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/auth/token": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Obtain a JWT access token (OAuth2 form)",
        "description": "OAuth2-compatible form-based login. Accepts `Content-Type: application/x-www-form-urlencoded` with `username` and `password` fields. Kept for backward compatibility and OpenAPI 'Authorize' button support. Prefer `POST /auth/login` for new clients.  Rate-limited via global middleware (100 req/min per user).",
        "operationId": "login_form_auth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_form_auth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Obtain a JWT access token (JSON body)",
        "description": "Preferred login endpoint. Accepts a JSON body `{username, password}`. Dispatches to LDAP when `LDAP_SERVER` is configured, otherwise uses local bcrypt authentication.",
        "operationId": "login_json_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login/local": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Local-only login",
        "description": "Authenticate with a local bcrypt password. Returns 403 when `LDAP_SERVER` is configured to prevent LDAP bypass. Use this endpoint in environments without LDAP.",
        "operationId": "login_local_auth_login_local_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get current user info",
        "description": "Return the username, org_id, and admin flag for the authenticated caller.",
        "operationId": "get_me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/users": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Create a local user",
        "description": "Create a new local user with a bcrypt-hashed password. The user is assigned to the caller's organization. Admin only.",
        "operationId": "create_user_auth_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/secrets": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "List secrets",
        "description": "Return metadata for all secrets in the caller's organization. Secret values are never included in the response. Admin only.",
        "operationId": "list_secrets_auth_secrets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SecretOut"
                  },
                  "type": "array",
                  "title": "Response List Secrets Auth Secrets Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Create a secret",
        "description": "Create a new named secret for the caller's organization. For `db` type secrets the plaintext `value` is encrypted (Fernet/AES) before storage. The value is **never** returned by any API endpoint. Admin only.",
        "operationId": "create_secret_auth_secrets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecretCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/secrets/{secret_id}": {
      "delete": {
        "tags": [
          "Authentication"
        ],
        "summary": "Delete a secret",
        "description": "Permanently delete a secret by ID. Admin only.",
        "operationId": "delete_secret_auth_secrets__secret_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "secret_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Secret Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_login_form_auth_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "password"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_form_auth_token_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "MeOut": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "org_id": {
            "type": "integer",
            "title": "Org Id"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          }
        },
        "type": "object",
        "required": [
          "username",
          "org_id",
          "is_admin"
        ],
        "title": "MeOut"
      },
...
      "SecretOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "secret_type": {
            "$ref": "#/components/schemas/SecretType"
          },
          "vault_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vault Path"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "name",
          "secret_type",
          "vault_path",
          "description",
          "created_at"
        ],
        "title": "SecretOut",
        "description": "Secret metadata \u2014 the plaintext value is never returned.",
        "examples": [
          {
            "created_at": "2024-01-15T09:00:00Z",
            "description": "NetBox API token for data source enrichment",
            "id": 3,
            "name": "netbox_api",
            "organization_id": 1,
            "secret_type": "db"
          }
        ]
      },
      "SecretType": {
        "type": "string",
        "enum": [
          "env",
          "vault",
          "db"
        ],
        "title": "SecretType"
      },
...
      "TokenOut": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "default": 28800
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenOut"
      },
      "UserCreate": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          },
          "is_ldap": {
            "type": "boolean",
            "title": "Is Ldap"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "is_admin",
          "is_ldap",
          "organization_id",
          "created_at"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}